requirements.txt
google-generativeai>=0.8.5
pypdf>=5.0.1
numpy>=2.0.0
streamlit>=1.39.0
python-dotenv>=1.0.1